Update parts list ---- read full commit message
Wen't a bit crazy on the parts list, feel free to be mad if i deleted something important, add it back if u like. I tried to force a format of 
Bold title + links
->>> then 
a list with 1)description + photo 2)quantity and then 3/4/5/6/7/etc all the random tidbits after that.

I kinda ran out of energy at the end.